+++ packages/connect-wallet-modal/src/hooks/useConnectWithLoading.ts
@@ -6,7 +6,7 @@ type UseConnectParams = Parameters<typeof useConnect>[0];
 export function useConnectWithLoading(params?: UseConnectParams) {
   const { setLoadingWalletId } = useReefKnotContext();
 
-  const { connectAsync } = useConnect(params);
+  const { mutateAsync: connectAsync } = useConnect(params);
 
   const connectWithLoading = async (
     walletId: string,