+++ packages/eslint-config-custom/index.js
@@ -9,7 +9,7 @@ module.exports = {
     node: true,
   },
   extends: ['turbo', '@lidofinance', 'prettier'],
-  plugins: ['prettier', '@typescript-eslint', 'react'],
+  plugins: ['prettier', '@typescript-eslint', 'react', 'react-hooks'],
   parser: '@typescript-eslint/parser',
   parserOptions: {
     ecmaFeatures: {
@@ -19,7 +19,7 @@ module.exports = {
     sourceType: 'module',
     project: project,
   },
-  ignorePatterns: ['*.js', 'dist/', 'node_modules/'],
+  ignorePatterns: ['*.js', '*.config.ts', 'dist/', 'node_modules/'],
   settings: {
     react: {
       version: 'detect',
@@ -32,5 +32,6 @@ module.exports = {
   },
   rules: {
     'prettier/prettier': ['error', {}, { usePrettierrc: true }],
+    'react/react-in-jsx-scope': 'off',
   },
 };