+++ packages/wallets/ctrl/src/icons/index.ts
@@ -1,1 +1,0 @@
-export { default as WalletIcon } from './ctrl.svg';