+++ packages/wallets/dappBrowserInjected/src/index.ts
@@ -1,3 +1,4 @@
+import 'viem/window';
 import { WalletAdapterType } from '@reef-knot/types';
 import { injected } from 'wagmi/connectors';
 import { isMobileOrTablet } from '@reef-knot/wallets-helpers';