+++ playwright-tests/config/wallet.config.ts
@@ -1,7 +1,6 @@
 import {
-  METAMASK_COMMON_CONFIG,
-  OKX_COMMON_CONFIG,
   CommonWalletConfig,
+  METAMASK_STABLE_COMMON_CONFIG,
 } from '@lidofinance/wallets-testing-wallets';
 
 export interface Wallet {
@@ -16,19 +15,11 @@ export const WALLETS = new Map<string, Wallet>([
     {
       name: 'metamask',
       config: {
-        ...METAMASK_COMMON_CONFIG,
+        ...METAMASK_STABLE_COMMON_CONFIG,
         LATEST_STABLE_DOWNLOAD_LINK:
           'https://github.com/MetaMask/metamask-extension/releases/download/v12.10.4/metamask-chrome-12.10.4.zip',
       },
       connectWalletEvent: 'metaMask connected',
-    },
-  ],
-  [
-    'okx',
-    {
-      name: 'okx',
-      config: OKX_COMMON_CONFIG,
-      connectWalletEvent: 'okx connected',
     },
   ],
 ]);