+++ playwright-tests/pages/components/index.ts
@@ -5,3 +5,4 @@ export * from './wallet-modal';
 export * from './toast';
 export * from './stake-block';
 export * from './wrap-unwrap-block';
+export * from './withdraw-block';